#1586 Add Csla.Blazor.WebAssembly project/package
