+++ NuGet/Definition/Csla.Blazor.WebAssembly.NuSpec
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <package>
   <metadata>
-    <id>Csla.Blazor</id>
-    <title>CSLA .NET - Blazor UI support</title>
-    <tags>CSLA Blazor aspnetcore</tags>
+    <id>Csla.Blazor.WebAssembly</id>
+    <title>CSLA .NET - Blazor WebAssembly support</title>
+    <tags>CSLA Blazor WebAssembly</tags>
     <version>4.6.3-Beta10</version>
     <authors>Marimer LLC</authors>
     <license type="expression">MIT</license>
@@ -12,26 +12,25 @@
     <iconUrl>https://raw.githubusercontent.com/MarimerLLC/csla/master/Support/Logos/csla.png</iconUrl>
     <requireLicenseAcceptance>true</requireLicenseAcceptance>
     <description>
-Provides UI support for Blazor.
+Provides UI support for client-side Blazor WebAssembly.
 
 CSLA .NET provides a home for your business logic. It is an application development framework that reduces the cost of building and maintaining applications. The framework enables developers to build an object-oriented business layer for their application that encapsulates all business, authorization and validation logic for the application.
     </description>
     <summary>
-Provides UI support for Blazor.
+Provides UI support for client-side Blazor WebAssembly.
     </summary>
     <language>en-US</language>
     <dependencies>
       <group targetFramework="netstandard2.1">
         <dependency id="Csla" version="[4.6.3-Beta10]" />
-        <dependency id="Microsoft.AspNetCore.Components" version="3.1.3" />
-        <dependency id="Microsoft.AspNetCore.Components.Web" version="3.1.3" />
-        <dependency id="Microsoft.AspNetCore.Components.Authorization" version="3.1.3" />
+        <dependency id="Csla.Blazor" version="[4.6.3-Beta10]" />
+        <dependency id="Microsoft.AspNetCore.Components.WebAssembly" version="3.2.0-preview4.20210.8" />
       </group>
     </dependencies>
   </metadata>
   <files>
     <!-- NetStandard Assembly -->
-    <file src="..\..\bin\Release\netstandard\netstandard2.1\**\Csla.Blazor.*" target="lib\netstandard2.1" />
+    <file src="..\..\bin\Release\netstandard\netstandard2.1\**\Csla.Blazor.WebAssembly.*" target="lib\netstandard2.1" />
     <!-- icon -->
     <file src="..\..\Support\Logos\csla.png" target="images\" />
   </files>
