+++ Source/Csla.Blazor.WebAssembly/Csla.Blazor.WebAssembly.csproj
@@ -22,7 +22,7 @@
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
     <OutputPath>..\..\Bin\Release\netstandard\</OutputPath>
-    <DocumentationFile>..\..\Bin\Release\netstandard\netstandard2.1\Csla.Blazor.xml</DocumentationFile>
+    <DocumentationFile>..\..\Bin\Release\netstandard\netstandard2.1\Csla.Blazor.WebAssembly.xml</DocumentationFile>
     <DefineConstants>TRACE;BLAZOR</DefineConstants>
   </PropertyGroup>
 
@@ -31,9 +31,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.Authorization" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="3.2.0-preview5.20216.8" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -41,6 +39,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Csla.Blazor\Csla.Blazor.csproj" />
     <ProjectReference Include="..\Csla.NetStandard2.0\Csla.NetStandard2.0.csproj" />
   </ItemGroup>
