+++ Source/Csla.Blazor.WebAssembly/Properties/AssemblyInfo.cs
@@ -5,7 +5,7 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("CSLA .NET Blazor Support")]
+[assembly: AssemblyTitle("CSLA .NET Blazor WebAssembly Support")]
 [assembly: AssemblyDescription("CSLA .NET Framework (NetStandard 2.1)")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Marimer LLC")]
